% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample of sparse microbiome count data}
\format{
The dataset contains 16 columns, indexed as A1 ~ A4, B1 ~ B4, C1 ~ C4, D1 ~ D4 to represent 4 treatment groups.
}
\usage{
sample_data
}
\description{
A sample data matrix with 100 features in 2 true clusters, 4 treatment groups with 4 replicates in each group.
}
\examples{
head(sample_data)
}
\keyword{datasets}
